password For User Uncomplete

+++ clinical/Pages/adminSettingsSecondPage.xaml.cs
@@ -117,30 +117,45 @@ namespace clinical.Pages
 
         private void delete(object sender, RoutedEventArgs e)
         {
-            if(packagesDataGrid.SelectedItem != null)
+            if (packagesDataGrid.SelectedItem != null)
             {
-                Package pack = (Package)packagesDataGrid.SelectedItem;
-                DB.DeletePackage(pack.PackageID);
-                packagesDataGrid.ItemsSource = DB.GetAllPackages();
-
+                MessageBoxResult result = MessageBox.Show("Are you sure you want to delete this Package?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+                    Package pack = (Package)packagesDataGrid.SelectedItem;
+                    DB.DeletePackage(pack.PackageID);
+                    packagesDataGrid.ItemsSource = DB.GetAllPackages();
+                }
             }
             else if (roomsDataGrid.SelectedItem != null)
             {
-                Room room = (Room)roomsDataGrid.SelectedItem;
-                DB.DeleteRoom(room.RoomID);
-                roomsDataGrid.ItemsSource = DB.GetAllRooms();
+                MessageBoxResult result = MessageBox.Show("Are you sure you want to delete this Room?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+                    Room room = (Room)roomsDataGrid.SelectedItem;
+                    DB.DeleteRoom(room.RoomID);
+                    roomsDataGrid.ItemsSource = DB.GetAllRooms();
+                }
             }
             else if (equipmentDataGrid.SelectedItem != null)
             {
-                Equipment eq = (Equipment)equipmentDataGrid.SelectedItem;
-                DB.DeleteEquipment(eq.EquipmentID);
-                equipmentDataGrid.ItemsSource = DB.GetAllEquipment();
+                MessageBoxResult result = MessageBox.Show("Are you sure you want to delete this Equipment Record?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+                    Equipment eq = (Equipment)equipmentDataGrid.SelectedItem;
+                    DB.DeleteEquipment(eq.EquipmentID);
+                    equipmentDataGrid.ItemsSource = DB.GetAllEquipment();
+                }
             }
             else if (appointmentTypesDataGrid.SelectedItem != null)
             {
-                AppointmentType ap = (AppointmentType)appointmentTypesDataGrid.SelectedItem;
-                DB.DeleteAppointmentType(ap.TypeID);
-                appointmentTypesDataGrid.ItemsSource = DB.GetAllAppointmentTypes();
+                MessageBoxResult result = MessageBox.Show("Are you sure you want to delete this appointment type?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+                    AppointmentType ap = (AppointmentType)appointmentTypesDataGrid.SelectedItem;
+                    DB.DeleteAppointmentType(ap.TypeID);
+                    appointmentTypesDataGrid.ItemsSource = DB.GetAllAppointmentTypes();
+                }
             }
         }
     }
